+++ Reach/test/utopia/reach/test/interactive/drawable/DrawableMouseTest.scala
@@ -16,13 +16,8 @@ import utopia.paradigm.color.{Color, Hsl}
 import utopia.paradigm.enumeration.RotationDirection.{Clockwise, Counterclockwise}
 import utopia.paradigm.shape.shape2d.area.Circle
 import utopia.paradigm.shape.shape2d.area.polygon.c4.bounds.Bounds
-import utopia.paradigm.shape.shape2d.insets.Insets
 import utopia.paradigm.shape.shape2d.vector.point.Point
-import utopia.paradigm.shape.shape2d.vector.size.Size
 import utopia.paradigm.transform.Adjustment
-import utopia.reach.component.label.drawable.DrawableCanvas
-import utopia.reach.cursor.DragTo
-import utopia.reach.window.ReachWindow
 
 /**
   * Tests mouse-listening and drawable canvas in Reach context
@@ -31,26 +26,14 @@ import utopia.reach.window.ReachWindow
   */
 object DrawableMouseTest extends App
 {
-	import utopia.reach.test.ReachTestContext._
+	import DrawableReachTestContext._
 	
-	private val viewBoundsPointer = Fixed(Bounds(0, 0, 640, 480))
-	
-	private val window = ReachWindow.contentContextual.borderless.using(DrawableCanvas) { (_, canvasF) =>
-		canvasF.withMinSize(Size.square(320))(Fixed(Bounds(0, 0, 640, 480)))
-	}
-	private val canvas = window.content
-	DragTo.resize.applyTo(window.content, Insets.symmetric(16))
-	
-	window.setToCloseOnEsc()
-	window.setToExitOnClose()
-	
-	canvas.viewHandlers += TestItem
-	canvas.viewHandlers += DrawBoundsDrawer
+	private val viewBoundsPointer = Fixed(viewBounds)
 	
 	println(canvas.stackSize)
 	
 	window.display(centerOnParent = true)
-	start()
+	start(TestItem, DrawBoundsDrawer)
 	
 	private object DrawBoundsDrawer extends AbstractDrawable(Foreground)
 	{